improved dual column layout
I added a true sidebar so elements on the right will float
independently of elements on the left. Changed .left to .post as we
transition to the sidebar layout. Also made some minor margin
improvements and removed unused selectors. Gave #overlay a z-index of 1
so the #content blurring will not glow distractingly. Major update but
will still require some tweaks that I can add soon.